Update mobile-services-filter-user-results-dotnet-backend.md
If you put the code after "TodoItem current = await InsertAsync(item);"
The UserId column in the database will be NULL.
So I think it should be "4. Locate the **PostTodoItem** method and add the following code at the end right before the **TodoItem current = await InsertAsync(item);** statement:" instead of "4. Locate the PostTodoItem method and add the following code at the end right before the return statement:"